+++ addons/yat/src/commands/builtin/Ping.cs
@@ -13,13 +13,11 @@ namespace YAT.Commands
 	)]
 	[Threaded]
 	[Argument("target", "string", "The target to ping.")]
-	[Options(
-		"-timeout=int(120:1200)",
-		"-delay=int(1:10)",
-		"-bytes=int(16:65500)",
-		"-ttl=int(1:255)",
-		"-fragment"
-	)]
+	[Option("-timeout", "int(120:1200)", "The timeout in milliseconds.")]
+	[Option("-delay", "int(1:10)", "The delay between pings in seconds.")]
+	[Option("-bytes", "int(16:65500)", "The number of bytes to send.")]
+	[Option("-ttl", "int(1:255)", "The time to live.")]
+	[Option("-fragment", null, "Fragment the packet.")]
 	public sealed class Ping : ICommand
 	{
 		public YAT Yat { get; set; }