+++ addons/yat/src/commands/builtin/QuickCommands.cs
@@ -8,7 +8,8 @@ namespace YAT.Commands
 {
 	[Command("quickcommands", "Manages Quick Commands.", "[b]Usage[/b]: quickcommands [i]action[/i] [i]name[/i] [i]command[/i]", "qc")]
 	[Argument("action", "[add, remove, list]", "The action to perform.")]
-	[Options("-name=string", "-command=string")]
+	[Option("-name", "string", "The name of the quick command.")]
+	[Option("-command", "string", "The command to execute when the quick command is called.")]
 	public sealed class QuickCommands : ICommand
 	{
 		public YAT Yat { get; set; }